requirements.txt
pytest==7.2.0
requests
voluptuous
pytest-voluptuous
allure-pytest
python-dotenv
